refactor: SalesRankingRedisRepository interface 제거
